Add AMRReductions class (#130)
* Add AMRReductions class

This class provides a user-friendly [and somewhat object oriented]
interface to Chombo's computeSum and computeNorm functions.

* Add new [Matter]Constraints classes

This is similar to the old Constraints classes but has a new
constructor which allows specifying the vars for
* Ham
* Moms
* Ham_abs_terms
* Moms_abs_terms
where the Moms are intervals. If Ham... is < 0, then it is not saved. If
the size of the Moms... intervals is 0, then it is not saved, if it is
1, then the pointwise norm is stored there and if it is 3 then it is
stored as the usual 3-vector.

As an example the ScalarField example has been modified to only have a
single c_Mom var.

* Fix for Tiago's comments on PR (#130)

* Modify new MatterConstraints constructor to make G_Newton an earlier
argument
* Calculate h_UU and chris outside of constraint_equations function and
pass via const reference to avoid calculating twice
* Fixed a missing newline character in ChomboParameters::read_params
output

* Add volume normalization ability to AMRReductions::norm

This divides the L^p norm by (vol)^(1/p) so that the value returned is
normalized by volume. By default this is not enabled.