Add Github action to run tests with Intel compiler
* The Intel compiler version used is whatever the latest "classic"
version is in the Intel apt software repositories (currently
2021.1.1-189).
* Fix GRAMRLevel.cpp includes to allow compiling without HDF5.
* A new minimal Make.defs.local example for the Intel compiler has been
added for this action.
* Since there is no suitable HDF5 package for the Intel compiler in the
Ubuntu repositories, USE_HDF is set to false in the Make.defs.local.
* Also update GCC action to use Ubuntu 20.04 github runner and remove
requesting the installation of some packages that are pre-installed.
* Both of these actions are set to run on every commit rather than just
on PRs (I think we have more than enough action minutes to be able to
do this).