Rewrite GRLevelData::plus with OpenMP and SIMD
Note that the prototype has also changed so that this function now only
loops over the `Box`es in the passed `DisjointBoxLayout` rather than the
intersection of the two `FArrayBox`es' `Box`es. This saves a little in
not looping through the interior ghost cells but in
`GRAMRLevel::updateODE`, m_grown_grids is passed which includes the
outer boundary ghosts in the case of Sommerfeld BCs.