Template CCZ4 and MatterCCZ4 classes over gauge and deriv
For backwards compatibility, they have been renamed CCZ4RHS and
MatterCCZ4RHS respectively. The old gauge has been moved into a new
MovingPuncturePlusGauge class and the old names, CCZ4 and MatterCCZ4
are defined using this. This resolves #146. Note that the checks pass,
in particular the BSSNTest and the CCZ4Test confirming that this
calculation is the same as before (at least for the specific
parameters).