Add sixth order spatial derivatives (#162)
* Add sixth order [spatial] derivatives

Also modify DerivativesUnitTest accordingly to test these.

* Add printing of value for parameters set to default

This requires adding a custom type_trait struct to check whether a type
is a std::array/std::vector or not which is done in ArrayTools. Then
SFINAE is used to choose the correct overload for printing the default
message.

* Add sixth order spatial derivatives to all examples

Add a max_spatial_derivative_order parameter to choose whether to use
4th or 6th order. The "max" is in the name as it only affects the
CCZ4RHS compute class but of course a user is free to use the max for
all their classes.

* Fix ChomboParameter comments

* Refactor calculation of default center in ChomboParameters

Now that default parameters are printed out when used, there is no need
to print out the default center separately.