Add SVE intrinsics
Unfortunately this does not use the vector length agnostic approach that
SVE was intended for as in this case, it would be much harder to modify
the existing simd infrastructure to accommodate passing the predicate
variable.

Note that this uses the optional __ARM_FEATURE_SVE_BITS extension to the
ACLE which may need to be manually enabled with a compiler flag
e.g. -msve-vector_bits=... with GCC v8.1 or greater.