Add open methods up to 4th order and SphericalExtractionUniform class
Also added a Test to check convergence of all methods.
This started as a test to see if a uniform distribution in the theta direction
of spherical coordinates would help at all in accuracy (since the usual ones
focuses a lot of points around the poles). It doesn't, but since it uses Open
methods, it has no point on the poles themselves, avoiding "nan"s that sometimes
appear when doing computations (it happened when computing the spin of the
AHFinder).