Improve BSSN formula to match CCZ4 version, assuming constraint satisfaction
In the new CCZ4 version of the formulas for Eij and Bij, these are explicitly symmetric and trace-free tensors (just from looking at the expression). For BSSN, the formulas do not naturally show this, yet it can be proved if assuming the constraints are satisfied (see Alcubierre chapter 8.3, from eq. 8.3.15 onwards). But since in BSSN the constraints are assumed to be 0, we can add precisely a certain multiple of the constrains (of the Hamiltonian to Eij and of the Momentum to Bij) to exchange matter terms by geometry terms that make the expression similar to the CCZ4 expression (without any Theta terms) and restoring the properties of being symmetric and trace-free, which is enforced numerically.

As additional notes, this change makes Bij have no matter terms, and also makes it explicit that Eij does NOT depend on the cosmological constant or the energy density even if non-zero (only on Sij).

This change was approved in discussions between @mirenradia, @KAClough, @JCAurre and myself.