Add Cubic Spline related classes
Cubic Spline is an interpolator that is quick and useful for 1D data. The Lagrange class could be used for interpolating 1D data, but Cubic Spline allows easily for non-uniform spacing 1D grid. This then can be used to interpolate data that exist in params.txt via the FileInterpolator1D. For instance, I have used this to solve 1D constraints in spherical symmetry elsewhere and then importing the data to the 3D grid.

It requires a matrix inversion, and that is done by a LU solver with the TriDiagonalMatrix class.

Two tests added. CubicSplineInterpolatorTest, to test the cubic spline interpolation, and FileInterpolator1DTest, to test the importing of file data to interpolate. The later can be also useful as an example of how to use these classes.

If "N", "x" and "y" are set in a separate file 'fileWithData.txt', one easy way to import them to the 'params.txt' without changing the file is running with the command:
'Main.ex params.txt FILE=fileWithData.txt'