Attempt to fix building of VTK grid
Note that now the vtkAMRBox indices are (0,0,0) where the lower
corner of the ghosted domain on the coarsest level is e.g. (-3,-3,-3)
in Chombo indices.
In theory this might make it annoying to calculate coordinates, but
these should be calculated from the associated vtkUniformGrid objects
anyway which have their origin at the correct place (i.e. Chombo's
(0,0,0)).
This is necessary as VTK expects the "origin" of the vtkAMRBoxes to be
the lower bound.