Make VTK/Catalyst log to a pout file
This reduces VTK log output to stderr and instead sends it to a
catalyst_pout (this name can be changed with the catalyst_pout_prefix
parameter) file (for each rank) under the usual pout_subpath like the
normal Chombo output (in the case MPI=TRUE). There is also a
catalyst_verbosity parameter which controls both the verbosity of the
VTK log output and the messages from CatalystAdaptor in the normal pout
file.

The -rdynamic flag will also be passed to the linker for the insitu
build target so that the VTK log stack trace (which is output in the
case of some termination signal) has more information for the [GR]Chombo
part of the stack.