Update fawcett Make.defs.local and modules
This is due to the new modules environment as of May 2022. Note that the
separation between KNL and Skylake Make.defs.local has been removed and
the new Make.defs.local should work for both. Furthermore, it has been
switched to using the new Intel oneAPI LLVM C++ compiler (icpx) rather
than continue to using the Intel C++ Classic compiler (icpc) which will
be deprecated in the next few years. Finally, support for the AHFinder
is enabled by default (a suitable PETSc module is loaded, linked with
and the cpp flag `-DUSE_AHFINDER` is used).