Fix slice visualization artifacts
The key step is to apply vtkParallelAMRUtilities::StripGhostLayers
(which removes layers of fine ghost cells that only partially cover
coarse cells) and vtkParallelAMRUtilties::BlankCells (which blanks out
coarser cells that are covered by finer cells on a different rank -- I
thought this was automatic previously) to the vtkOverlappingAMR object
(this is done in the [VisIt Chombo
Reader](https://gitlab.kitware.com/paraview/visitbridge/-/blob/df098f4148a96d62c388861c1d476039e02224ae/databases/AvtAlgorithms/vtkAvtSTMDFileFormatAlgorithm.cxx#L511-L514)).
Since the former might require deep copying (in case stripping happens),
a temporary vtkOverlappingAMR object is instantiated in
CatalystAdaptor::coprocess which is used to create the grid and add the
fields. This is then copied (either shallow or deep as necessary) to the
m_vtk_grid_ptr member which is passed to Catalyst. Note that now, no
cells are manually marked as ghosts in CatalystAdaptor::build_vtk_grid
as I observed gaps when doing so. The only marking of ghosts now comes
from BlankCells.