Added symmetry correction flag and symmetry factor variable
1. The symmetry correction flag is read in ChomboParameters. It is
defaulted to "off" if the parameter is not provided.
2. If symmetry correction is flag is on, the symmetry factor is
calculated based off of the number of hi or lo boundaries chosen
to be reflective (choice 2).
3. The symmetry factor is a chombo parameter, and must be given to
any AMRReduction function which integrates over the entire box
(currently norm and sum). The symmetry factor is defaulted to 1
if it is not provided to the AMRReduction scheme.
4. The result of sum or norm is multiplied by the symmetry factor
before it is returned, in the case of a single variable.

The symmetry factor is provided correctly to AMRReductions for
the BinaryBH example ONLY in this commit.