Symmetry factor implementation
Created a series of subroutines to directly read in
the symmetry factor associated with using reflective
boundary conditions in the GRAMR class, which can then be
drawn out for use by the sum() and norm() AMRReductions
functions.

1. If the symmetry_correction flag in params is turn on,
the symmetry factor is calculated in BoundaryConditions
as a parameter, then assigned to a private BoundaryConditions
variable.
2. A GRAMRLevel subroutine, get_symm(), calls the BoundaryConditions
subroutine to return the private symmetry factor variable.
3. AMRReductions reads in the symmetry factor from get_gramrlevels
in the same style as m_coarsest_dx.
4. The symmetry factor is applied to the norm() and sum() results
immediately before they are returned.