Bug fix for NanCheck exit inconsistency
This commit addresses Issue 238 primarily in the NanCheck.cpp file,
where the switch to the master thread (#pragma omp master) is placed
AFTER the start of the "if" statement involving the "stop" variable,
rather than before.

I suspect that since the "stop" variable was altered under the "atomic
write" OpenMP command, if any thread other than the master thread caught
the Nan then the memory storing the "stop" variable would not be shared
with master, and thus the if statement leading to the error trap would
not be tripped if evaluated on the master thread. This change
effectively moves the evaluation of that logic statement to each OpenMP
thread individually.

This commit also has an error-trip built into the ScalarField example,
where in InitialScalarData I set the value of the extrinsic curvature K
somewhere in the box to Nan. This version has a commented-out NanCheck
right after the initial data as well, which can be used to debug the
current issue (see end).

I also added the specificPostTimeStep function to the ScalarField
example, and moved the NanCheck that was called in specificAdvance to
this function.

PS. Another way to solve this problem could be to transfer the "stop"
variable to the master thread in some way before the logic statement.
This may address certain error output issues that are present in the
current commit, meaning that if the error is caught on a non-master
thread it is also printed by that same thread. We may need to look into
this in the future.